+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 - Show entities whose record has been deleted instead of raising in the UI
+- Add `Flipside::Flippable` to remove a record's entities when it is destroyed
 
 ## [0.3.4] - 2026-03-04
 - Support an array of default objects